relocate `CheckDACertStatus`
